got rid of one coredump (size_t < 0) and two warnings in the buffer code

+++ mediatomb/src/buffered_io_handler.cc
@@ -122,7 +122,7 @@ void BufferedIOHandler::threadProc()
                     int currentFillSize = b - a;
                     if (currentFillSize <= 0)
                         currentFillSize += bufSize;
-                    if (currentFillSize >= initialFillSize)
+                    if ((size_t)currentFillSize >= initialFillSize)
                     {
                         log_debug("buffer: initial fillsize reached\n");
                         waitForInitialFillSize = false;
